refactor(api/user): move IUserStore to User folder
