+++ modules/back-end/src/Infrastructure/Users/IUserStore.cs
@@ -1,6 +1,6 @@
 using Domain.Users;
 
-namespace Infrastructure.Identity;
+namespace Infrastructure.Users;
 
 public interface IUserStore
 {
