+++ modules/back-end/src/Infrastructure/Users/MongoDbUserStore.cs
@@ -3,7 +3,7 @@ using Infrastructure.MongoDb;
 using MongoDB.Driver;
 using MongoDB.Driver.Linq;
 
-namespace Infrastructure.Identity;
+namespace Infrastructure.Users;
 
 public class MongoDbUserStore : IUserStore
 {
